requirements.txt
attrs==22.1.0
autopep8==2.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Flask==2.2.2
Flask-MySQLdb==1.0.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==22.0
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.0
tomli==2.0.1
Werkzeug==2.2.2
